third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/popup-same-origin-with-same-site.https-expected.txt

This is a testharness.js-based test.
FAIL same-origin document opening popup to https://www1.web-platform.test:8444 with COOP: "" assert_equals: Openee browsing context closed expected true but got false
FAIL same-origin document opening popup to https://www1.web-platform.test:8444 with COOP: "unsafe-none" assert_equals: Openee browsing context closed expected true but got false
FAIL same-origin document opening popup to https://www1.web-platform.test:8444 with COOP: "jibberish" assert_equals: Openee browsing context closed expected true but got false
FAIL same-origin document opening popup to https://www1.web-platform.test:8444 with COOP: "same-site" assert_equals: Openee browsing context closed expected true but got false
FAIL same-origin document opening popup to https://www1.web-platform.test:8444 with COOP: "same-site unsafe-allow-outgoing" assert_equals: Openee browsing context closed expected true but got false
FAIL same-origin document opening popup to https://www1.web-platform.test:8444 with COOP: "same-origin unsafe-allow-outgoing" assert_equals: Openee browsing context closed expected true but got false
FAIL same-origin document opening popup to https://www1.web-platform.test:8444 with COOP: "same-origin" assert_equals: Openee browsing context closed expected true but got false
FAIL same-origin document opening popup to https://www1.web-platform.test:8444 with COOP: "same-origin-allow-popups" assert_equals: Openee browsing context closed expected true but got false
Harness: the test ran to completion.
